refactor: move sprite types
