+++ frontend/src/game/levels/list.ts
@@ -1,5 +1,6 @@
 import testLevel from './testLevel';
-import { LevelConfig, LevelId } from './types';
+import { LevelConfig } from './levelConfig';
+import { LevelId } from './types';
 
 const levelList:Record<LevelId, LevelConfig> = {
   [LevelId.test]: testLevel,
