+++ frontend/src/game/levels/types/index.ts
@@ -1,11 +1,7 @@
-import {
-  LevelConfig, LevelEntityConfig, LevelLoadingConfig, LevelSurfaceConfig, surfaceConfigList,
-} from '../levelConfig';
 import LevelId from './levelIds';
 import LevelLoad from './levelLoad';
 import SurfaceType from './surfaceType';
 
 export {
-  LevelConfig, LevelId, LevelLoad, SurfaceType, surfaceConfigList,
-  LevelEntityConfig, LevelLoadingConfig, LevelSurfaceConfig,
+  LevelId, LevelLoad, SurfaceType,
 };
