+++ frontend/src/game/sprites/index.ts
@@ -1,6 +1,11 @@
 import { Point, Rectangle } from '../../shapes';
-import SpriteConfig from './typeConfig';
-import SpriteFrame from './typeSpriteFrame';
+import SpriteConfig from './spriteConfig';
+
+type SpriteFrame = {
+  position:Rectangle;
+  positionReverse?:Rectangle;
+  ends:number;
+};
 
 class SpriteAnimation {
   private readonly imgSource:CanvasImageSource;
