--- frontend/src/game/sprites/typeSpriteFrame.ts
@@ -1,9 +1,0 @@
-import { Rectangle } from '../../shapes';
-
-type SpriteFrame = {
-  position:Rectangle;
-  positionReverse?:Rectangle;
-  ends:number;
-};
-
-export default SpriteFrame;
